devtools/client/shared/shim/test/test_service_wm.html

<!DOCTYPE html>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=1310279
-->
<head>
  <title>Test for Bug 1310279 - replace Services.wm</title>
  <script type="text/javascript" src="/MochiKit/MochiKit.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css"
        href="chrome://mochikit/content/tests/SimpleTest/test.css">

  <script type="application/javascript;version=1.8">
  "use strict";
  var exports = {}
  var module = {exports};
  </script>

  <script type="application/javascript;version=1.8"
	  src="resource://devtools/client/shared/shim/Services.js"></script>
</head>
<body>

<script type="application/javascript;version=1.8">
  "use strict";

  function hurray(window) {
    ok(true, "window loaded");
    SimpleTest.finish();
  }

  SimpleTest.waitForExplicitFinish();
  Services.wm.getMostRecentWindow().openUILinkIn("file_service_wm.html");

</script>
</body>
